Add support for scheduled and delayed jobs
Because jobs require persisent logbooks to function, we also stop
requiring the user to do all the bookkeeping externally
and handle it automatically inside the job board itself.

Some common functionality was factored out into base classes
to simplify job board implementation boilerplate. There's more
to do there, but it's a start.

Scheduled jobs work by posting a new job with the same details
after the job is consumed, if there is a next iteration remaining
on the schedule.

Delayed jobs are just jobs with a `run_at` timestamp that are
ignored until that time is no longer in the future.

We should probably move some of the fields from the job details dict
into top-level fields on the job to make them first-class citizens,
but that can wait.