Improve JSON-encoding failures to be more resilient
We missed the most common JSON encoding error TypeError,
but also OverflowError is possible according to the docs.

Moved to base Exception type with the fallback to guarantee
that it's covered, as I can't find docs or an example of
failing to encode something as six.text_type(), but I would
imagine the only case would be throwing an exception from a
__repr__ method, which could be anything. So to be safe, just
trap everything.

Fixes #39